backend/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
psycopg2-binary==2.9.6 # PostgreSQL database adapter
djangorestframework-simplejwt==5.3.0 # For JWT authentication
python-dotenv==1.0.0 # To load environment variables from .env file